refactor: change some file names
