+++ src/App.js
@@ -1,10 +1,10 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { useEffect, useState } from 'react';
 
-import Login from './pages/Login';
-import Cadastro from './pages/Cadastro';
+import LoginScreen from './pages/LoginScreen';
+import SignupScreen from './pages/SignupScreen';
 import HomeScreen from './pages/Home/HomeScreen';
-import Produto from './pages/Produto';
+import ProductScreen from './pages/ProductScreen';
 
 import UserContext from './contexts/UserContext';
 
@@ -33,9 +33,9 @@ function App() {
         <Header />
         <Routes>
           <Route path="/" element={<HomeScreen />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/cadastro" element={<Cadastro />} />
-          <Route path="/produto/:productId" element={<Produto />} />
+          <Route path="/login" element={<LoginScreen />} />
+          <Route path="/cadastro" element={<SignupScreen />} />
+          <Route path="/produto/:productId" element={<ProductScreen />} />
         </Routes>
       </BrowserRouter>
     </UserContext.Provider>
